+++ changelog.md
@@ -1,6 +1,7 @@
 ## Upcoming
 * Added a new Creative Removal feature
   * The `insanesurvivaloverhaul:creative_removal` item tag can be used to remove items from the creative inventory
+* Undead now burn in daylight even if there's a block above them
 * Enchanted items can no longer be used for anvil crafting
 
 ## 5.4.3.1